+++ docs/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 ---
+### 10 SEP 2025
+- Created multiple tests when `batchRename` throws the error `Provided path must be an absolute path to a directory or file`.
+---
 ### 09 SEP 2025
 - Created `tests` and `mock-files` subdirectories.
 - Installed `ts-jest` and `@types/jest` packages to `devDependencies`.